Move ErrorResponse to Api/Contracts — it's a response DTO, not infra
